requirements.txt
alembic==0.9.9
anyjson==0.3.3
beautifulsoup4==4.6.0
cffi==1.11.5
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.6
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
